src/requirements.txt
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-SocketIO==4.0.0
Flask-SQLAlchemy==2.4.1
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
oauthlib==3.1.0
requests-oauthlib==1.2.0
python-dotenv==0.10.3
google-auth==1.6.3